requirements.txt
astroid==2.2.5
beautifulsoup4==4.8.1
bleach==3.1.0
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.3.9
cffi==1.13.1
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.8
cycler==0.10.0
docutils==0.15.2
firebase-admin==2.17.0
Flask==1.0.3
geographiclib==1.50
geopy==1.20.0
google==2.0.2
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-core==1.0.3
google-cloud-firestore==1.2.0
google-cloud-storage==1.16.1
google-cloud-translate==1.7.0
google-cloud-vision==0.39.0
google-resumable-media==0.3.2
google-streetview==1.2.9
googleapis-common-protos==1.6.0
grpcio==1.21.1
httplib2==0.13.0
idna==2.8
isort==4.3.16
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
kwconfig==1.1.7
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
matplot==0.1.9
matplotlib==3.1.1
mccabe==0.6.1
msgpack==0.6.1
numpy==1.17.3
pandas==0.25.1
paypalrestsdk==1.13.1
Pillow==6.2.1
pkginfo==1.5.0.1
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
Pygments==2.4.2
pylint==2.3.1
pyloco==0.0.139
pyOpenSSL==19.0.0
pyparsing==2.4.2
python-dateutil==2.8.0
python-google-places==1.4.1
python-googlegeocoder==0.4.0
pytz==2019.1
readme-renderer==24.0
requests==2.21.0
requests-toolbelt==0.9.1
rsa==4.0
SimpleWebSocketServer==0.1.1
six==1.12.0
soupsieve==1.9.4
stripe==2.37.2
tqdm==4.36.1
twine==2.0.0
typed-ast==1.3.1
typing==3.7.4.1
uritemplate==3.0.0
urllib3==1.24.1
ushlex==0.99.1
virtualenv==16.6.1
watson-developer-cloud==2.10.1
webencodings==0.5.1
websocket-client==0.48.0
Werkzeug==0.15.4
wrapt==1.11.1
